backend/requirements.txt
pandas
requests
openpyxl
python-dotenv
scikit-learn
numpy
google-generativeai
httpx
supabase
python-jose[cryptography]
fastapi
uvicorn
